+++ plugins/woocommerce-admin/packages/explat/CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+-   Fix and refactor explat polling to use setTimeout #7274
+
 # 1.0.1
 
 -   Update ExPlat client dependencies